Split header and implementation files
