+++ src/search/search_encoding.h
@@ -114,7 +114,7 @@ struct SearchKey {
 
   void PutIndex(std::string *dst) const { PutSizedString(dst, index); }
 
-  void PutHnswLevelType(std::string *dst, HnswLevelType type) const { PutFixed8(dst, uint8_t(type)); }
+  static void PutHnswLevelType(std::string *dst, HnswLevelType type) { PutFixed8(dst, uint8_t(type)); }
 
   void PutHnswLevelPrefix(std::string *dst, uint16_t level) const {
     PutNamespace(dst);
@@ -413,15 +413,15 @@ struct HnswNodeFieldMetadata {
   uint16_t num_neighbours;
   std::vector<double> vector;
 
-  HnswNodeFieldMetadata() {}
+  HnswNodeFieldMetadata() = default;
   HnswNodeFieldMetadata(uint16_t num_neighbours, std::vector<double> vector)
-      : num_neighbours(num_neighbours), vector(vector) {}
+      : num_neighbours(num_neighbours), vector(std::move(vector)) {}
 
   void Encode(std::string *dst) const {
     PutFixed16(dst, uint16_t(num_neighbours));
     PutFixed16(dst, uint16_t(vector.size()));
-    for (size_t i = 0; i < vector.size(); ++i) {
-      PutDouble(dst, vector[i]);
+    for (auto element : vector) {
+      PutDouble(dst, element);
     }
   }
 
@@ -431,7 +431,7 @@ struct HnswNodeFieldMetadata {
     }
     GetFixed16(input, (uint16_t *)(&num_neighbours));
 
-    uint16_t dim;
+    uint16_t dim = 0;
     GetFixed16(input, (uint16_t *)(&dim));
 
     if (input->size() != dim * sizeof(double)) {
